+++ lib/cassandra/driver.rb
@@ -84,7 +84,7 @@ module Cassandra
                                      cluster_registry,
                                      cluster_schema,
                                      cluster_metadata,
-                                     load_balancing_policy,
+                                     profile_manager.default_profile.load_balancing_policy,
                                      reconnection_policy,
                                      address_resolution_policy,
                                      connector,
@@ -104,21 +104,22 @@ module Cassandra
                         cluster_metadata,
                         execution_options,
                         connection_options,
-                        load_balancing_policy,
+                        profile_manager,
                         reconnection_policy,
-                        retry_policy,
                         address_resolution_policy,
                         connector,
                         futures_factory,
-			timestamp_generator)
+			                  timestamp_generator)
     end
 
     let(:execution_options) do
-      Execution::Options.new(consistency: consistency,
+      Execution::Options.new(consistency: profile_manager.default_profile.consistency,
                              trace: trace,
                              page_size: page_size,
-                             timeout: timeout,
-                             idempotent: false)
+                             timeout: profile_manager.default_profile.timeout,
+                             idempotent: false,
+                             load_balancing_policy: profile_manager.default_profile.load_balancing_policy,
+                             retry_policy: profile_manager.default_profile.retry_policy)
     end
 
     let(:connection_options) do
@@ -169,7 +170,7 @@ module Cassandra
     let(:page_size)                 { 10000 }
     let(:heartbeat_interval)        { 30 }
     let(:idle_timeout)              { 60 }
-    let(:timeout)                   { 12 }
+    let(:timeout)                   { Cassandra::Execution::Profile::DEFAULT_TIMEOUT }
     let(:synchronize_schema)        { true }
     let(:schema_refresh_delay)      { 1 }
     let(:schema_refresh_timeout)    { 10 }
@@ -180,7 +181,14 @@ module Cassandra
     let(:connections_per_local_node)  { nil }
     let(:connections_per_remote_node) { nil }
     let(:requests_per_connection) { nil }
-
+    let(:default_execution_profile) {
+      Cassandra::Execution::Profile.new(load_balancing_policy: load_balancing_policy,
+                                        retry_policy: retry_policy,
+                                        consistency: consistency,
+                                        timeout: timeout)
+    }
+    let(:execution_profiles) { {} }
+    let(:profile_manager) { Cassandra::Execution::ProfileManager.new(default_execution_profile, execution_profiles) }
     let(:listeners) { [] }
 
     def initialize(defaults = {})
@@ -188,8 +196,11 @@ module Cassandra
     end
 
     def connect(addresses)
-      load_balancing_policy.setup(cluster)
-      cluster_registry.add_listener(load_balancing_policy)
+      profile_manager.load_balancing_policies.each do |lbp|
+        lbp.setup(cluster)
+        cluster_registry.add_listener(lbp)
+      end
+
       cluster_registry.add_listener(control_connection)
       listeners.each do |listener|
         cluster.register(listener)