+++ spec/cassandra/execution/options_spec.rb
@@ -21,7 +21,45 @@ require 'spec_helper'
 module Cassandra
   module Execution
     describe(Options) do
-      let(:base_options) { Options.new(timeout: 10, consistency: :one) }
+      let(:load_balancing_policy) { double('lbp') }
+      let(:retry_policy) { double('retry_policy') }
+      let(:base_options) {
+        Options.new(timeout: 10, consistency: :one, load_balancing_policy: load_balancing_policy,
+                    retry_policy: retry_policy)
+      }
+      before do
+        allow(load_balancing_policy).to receive(:host_up)
+        allow(load_balancing_policy).to receive(:host_down)
+        allow(load_balancing_policy).to receive(:host_found)
+        allow(load_balancing_policy).to receive(:host_lost)
+        allow(load_balancing_policy).to receive(:setup)
+        allow(load_balancing_policy).to receive(:teardown)
+        allow(load_balancing_policy).to receive(:distance)
+        allow(load_balancing_policy).to receive(:plan)
+
+        allow(retry_policy).to receive(:read_timeout)
+        allow(retry_policy).to receive(:write_timeout)
+        allow(retry_policy).to receive(:unavailable)
+      end
+
+      context :initialize do
+        it 'should yell if load-balancing policy is invalid' do
+          ['junk', nil].each do |val|
+            expect {
+              Options.new(load_balancing_policy: val, retry_policy: retry_policy, consistency: :one)
+            }.to raise_error(ArgumentError)
+          end
+        end
+
+        it 'should yell if retry policy is invalid' do
+          ['junk', nil].each do |val|
+            expect {
+              Options.new(load_balancing_policy: load_balancing_policy, retry_policy: val, consistency: :one)
+            }.to raise_error(ArgumentError)
+          end
+        end
+      end
+
       it 'should allow nil timeout to override base non-nil timeout option' do
         result = Options.new({timeout: nil}, base_options)
         expect(result.timeout).to be_nil
@@ -35,6 +73,15 @@ module Cassandra
       it 'should not override base timeout if not specified' do
         result = Options.new({}, base_options)
         expect(result.timeout).to eq(10)
+      end
+
+      it 'should override with execution-profile and simple attribute' do
+        profile = Profile.new(timeout: 21, consistency: :quorum)
+        result = base_options.override(profile, timeout: 42)
+        expect(result.load_balancing_policy).to be(load_balancing_policy)
+        expect(result.retry_policy).to be(retry_policy)
+        expect(result.consistency).to be(:quorum)
+        expect(result.timeout).to eq(42)
       end
     end
   end